fix: extract ObxError to separate file and improve null safety
- Extract ObxError class to lib/get_common/obx_error.dart
- Make ObxError implement Exception to satisfy Dart linting rules
- Remove commented-out dead code in list_notifier.dart
- Add null safety assertions to addListenerId() and disposeId()
