+++ lib/get_common/obx_error.dart
@@ -1,0 +1,14 @@
+class ObxError implements Exception {
+  const ObxError();
+  @override
+  String toString() {
+    return """
+      [Get] the improper use of a GetX has been detected. 
+      You should only use GetX or Obx for the specific widget that will be updated.
+      If you are seeing this error, you probably did not insert any observable variables into GetX/Obx 
+      or insert them outside the scope that GetX considers suitable for an update 
+      (example: GetX => HeavyWidget => variableObservable).
+      If you need to update a parent widget and a child widget, wrap each one in an Obx/GetX.
+      """;
+  }
+}
