+++ lib/get_rx/src/rx_types/rx_core/rx_interface.dart
@@ -9,21 +9,10 @@ abstract class RxInterface<T> implements ValueListenable<T> {
   void close();
 
   /// Calls `callback` with current value, when the value changes.
-  StreamSubscription<T> listen(void Function(T event) onData,
-      {Function? onError, void Function()? onDone, bool? cancelOnError});
-}
-
-class ObxError {
-  const ObxError();
-  @override
-  String toString() {
-    return """
-      [Get] the improper use of a GetX has been detected. 
-      You should only use GetX or Obx for the specific widget that will be updated.
-      If you are seeing this error, you probably did not insert any observable variables into GetX/Obx 
-      or insert them outside the scope that GetX considers suitable for an update 
-      (example: GetX => HeavyWidget => variableObservable).
-      If you need to update a parent widget and a child widget, wrap each one in an Obx/GetX.
-      """;
-  }
+  StreamSubscription<T> listen(
+    void Function(T event) onData, {
+    Function? onError,
+    void Function()? onDone,
+    bool? cancelOnError,
+  });
 }
